Added libary providing basic compatibility for bitwise operations
love on desktop is using luajit 2.1 which has builtin support for bitwise ops.
Lua 5.1., which is currently used for building, love.js does not.

The following  worked with love on desktop but not with love.js:

        require("bit")
        x = bit.bxor(1,2)

Using the MIT-licenced module from https://bitop.luajit.org/ makes it possible to
use the same (efficient) code on desktop as well as web.

The easiest (and probably not the prettiest) way to include this was to add the bit.c file
directly to the lua-5.1-sources.